nf RDM-75 BigInteger вместо Integer, исправлены тесты. Проверка при изменении типа атрибута

See merge request platform/versioned-data-storage!6

+++ versioned-data-storage-pg-impl/src/main/java/ru/i_novus/platform/versioned_data_storage/pg_impl/model/IntegerField.java
@@ -4,11 +4,13 @@ import ru.i_novus.platform.datastorage.temporal.model.Field;
 import ru.i_novus.platform.datastorage.temporal.model.FieldValue;
 import ru.i_novus.platform.datastorage.temporal.model.value.IntegerFieldValue;
 
+import java.math.BigInteger;
+
 /**
  * @author lgalimova
  * @since 23.03.2018
  */
-public class IntegerField extends Field<Integer> {
+public class IntegerField extends Field<BigInteger> {
     public static final String TYPE = "bigint";
 
     public IntegerField(String name) {
@@ -21,7 +23,7 @@ public class IntegerField extends Field<Integer> {
     }
 
     @Override
-    public FieldValue valueOf(Integer value) {
+    public FieldValue valueOf(BigInteger value) {
         return new IntegerFieldValue(getName(), value);
     }
 }
